requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.0.4
django-phonenumber-field==7.3.0
gunicorn==21.2.0
packaging==24.0
phonenumbers==8.13.35
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
reportlab==4.2.0
sqlparse==0.5.0
tzdata==2024.1
whitenoise==6.6.0
